pyproject.toml
[project]
name = "sage"
version = "0.4.0"
requires-python = ">=3.12"
dependencies = [
    "stix2>=3.0.1",
    "pycti>=6.3.0",
    "google-cloud-spanner>=3.49.0",
    "google-cloud-storage>=2.18.0",
    "structlog>=24.4.0",
    "requests>=2.32.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "pip-audit>=2.7.0",
    "pyvis>=0.3.2",
    "httpx>=0.27.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
